refactor(header-chain): package root by domain
Keep the crate root focused on client-facing facades while preserving all flat public API paths.
