+++ crates/zakura-header-chain/src/identity/mod.rs
@@ -1,0 +1,9 @@
+//! Stable domain identities and monotonic durable counters.
+
+mod counters;
+mod keys;
+
+pub use counters::{
+    CounterExhausted, FinalityEpoch, HeaderGeneration, StateVersion, VerifiedGeneration,
+};
+pub use keys::{BranchId, EvidenceId, HeaderId, OperatorInvalidationId, SourceId};
